modified:   pom.xml 	modified:   src/main/java/com/bibliotheque/controller/EmpruntController.java 	modified:   src/main/java/com/bibliotheque/controller/LivreController.java 	modified:   src/main/java/com/bibliotheque/dao/EmpruntDAO.java 	modified:   src/main/java/com/bibliotheque/dao/impl/EmpruntDAOImpl.java 	modified:   src/main/java/com/bibliotheque/dao/impl/MembreDAOImpl.java 	modified:   src/main/java/com/bibliotheque/infra/DatabaseConnection.java 	modified:   src/main/java/com/bibliotheque/model/Emprunt.java 	modified:   src/main/java/com/bibliotheque/model/Livre.java 	modified:   src/main/java/com/bibliotheque/service/BibliothequeService.java 	modified:   src/main/java/com/bibliotheque/service/EmpruntService.java 	new file:   src/main/resources/db/migration/V1__create_livres_table.sql 	new file:   src/main/resources/db/migration/V2__create_membres_table.sql 	new file:   src/main/resources/db/migration/V3__create_emprunts_table.sql 	modified:   src/main/resources/fxml/LivreView.fxml

+++ src/main/java/com/bibliotheque/service/BibliothequeService.java
@@ -305,7 +305,7 @@ public class BibliothequeService {
     public List<String> historiqueEmprunts(Long membreId) {
         if (membreId == null) return Collections.emptyList();
 
-        List<Emprunt> emprunts = empruntDAO.findByMemberId(membreId.intValue());
+        List<Emprunt> emprunts = empruntDAO.findByMemberId(membreId);
         return emprunts.stream()
             .map(e -> "Emprunt du " + e.getDateEmprunt() + (e.getDateRetour() != null ? " retourné le " + e.getDateRetour() : " en cours"))
             .collect(java.util.stream.Collectors.toList());
